+++ pinot-controller/src/main/java/org/apache/pinot/controller/LeadControllerManager.java
@@ -18,37 +18,44 @@
  */
 package org.apache.pinot.controller;
 
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import org.apache.helix.model.ExternalView;
 import org.apache.pinot.common.config.TableNameBuilder;
 import org.apache.pinot.common.utils.helix.LeadControllerUtils;
 import org.apache.pinot.controller.helix.core.PinotHelixResourceManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.pinot.common.utils.CommonConstants.Helix.LEAD_CONTROLLER_RESOURCE_NAME;
 import static org.apache.pinot.common.utils.CommonConstants.Helix.NUMBER_OF_PARTITIONS_IN_LEAD_CONTROLLER_RESOURCE;
 
 
 public class LeadControllerManager {
   private static final Logger LOGGER = LoggerFactory.getLogger(LeadControllerManager.class);
 
-  private Map<Integer, Integer> _partitionIndexCache;
+  private Set<Integer> _partitionIndexCache;
   private PinotHelixResourceManager _pinotHelixResourceManager;
   private volatile boolean _amIHelixLeader = false;
 
   public LeadControllerManager() {
-    _partitionIndexCache = new ConcurrentHashMap<>();
+    _partitionIndexCache = ConcurrentHashMap.newKeySet();
   }
 
+  /**
+   * Registers {@link PinotHelixResourceManager} in LeadControllerManager, which is needed to get the external view of lead controller resource.
+   */
   public void registerResourceManager(PinotHelixResourceManager pinotHelixResourceManager) {
     _pinotHelixResourceManager = pinotHelixResourceManager;
   }
 
   /**
-   * Check whether the current controller is the leader for the given table. Return true if current controller is the leader for this table.
+   * Unregisters {@link PinotHelixResourceManager}.
+   */
+  public void unregisterResourceManager() {
+    _pinotHelixResourceManager = null;
+  }
+
+  /**
+   * Checks whether the current controller is the leader for the given table. Return true if current controller is the leader for this table.
    * Otherwise check whether the current controller is helix leader.
    * @param tableName table name with/without table type.
    */
@@ -56,7 +63,7 @@ public class LeadControllerManager {
     String rawTableName = TableNameBuilder.extractRawTableName(tableName);
     int partitionIndex =
         LeadControllerUtils.getPartitionIdForTable(rawTableName, NUMBER_OF_PARTITIONS_IN_LEAD_CONTROLLER_RESOURCE);
-    if (_partitionIndexCache.containsKey(partitionIndex)) {
+    if (_partitionIndexCache.contains(partitionIndex)) {
       return true;
     } else {
       return isHelixLeader();
@@ -64,35 +71,19 @@ public class LeadControllerManager {
   }
 
   /**
-   * Get lead controller id and partition index for given table. Returns -1 as partition index and null as lead controller id if not found.
-   * @param tableName table name with/without table type.
-   * @return lead controller id and partition index
+   * Given a partition name, marks current controller as lead controller for this partition by caching the partition index to current controller.
+   * @param partitionName partition name in lead controller resource, e.g. leadControllerResource_0.
    */
-  public LeadControllerResponse getLeadControllerAndPartitionIdForTable(String tableName) {
-    String rawTableName = TableNameBuilder.extractRawTableName(tableName);
-    ExternalView leadControllerResourceExternalView = _pinotHelixResourceManager.getHelixAdmin()
-        .getResourceExternalView(_pinotHelixResourceManager.getHelixClusterName(), LEAD_CONTROLLER_RESOURCE_NAME);
-    if (leadControllerResourceExternalView == null) {
-      return new LeadControllerResponse();
-    }
-    Set<String> partitionSet = leadControllerResourceExternalView.getPartitionSet();
-    if (partitionSet == null || partitionSet.isEmpty()) {
-      return new LeadControllerResponse();
-    }
-
-    int numPartitions = partitionSet.size();
-    int partitionIndex = LeadControllerUtils.getPartitionIdForTable(rawTableName, numPartitions);
-    String leadControllerId =
-        LeadControllerUtils.getLeadControllerForTable(leadControllerResourceExternalView, rawTableName);
-    return new LeadControllerResponse(partitionIndex, leadControllerId);
-  }
-
   public synchronized void addPartitionLeader(String partitionName) {
     LOGGER.info("Add Partition: {} to LeadControllerManager", partitionName);
     int partitionIndex = LeadControllerUtils.extractPartitionIndex(partitionName);
-    _partitionIndexCache.put(partitionIndex, partitionIndex);
+    _partitionIndexCache.add(partitionIndex);
   }
 
+  /**
+   * Given a partition name, removes current controller as lead controller for this partition by removing the partition index from current controller.
+   * @param partitionName partition name in lead controller resource, e.g. leadControllerResource_0.
+   */
   public synchronized void removePartitionLeader(String partitionName) {
     LOGGER.info("Remove Partition: {} from LeadControllerManager", partitionName);
     int partitionIndex = LeadControllerUtils.extractPartitionIndex(partitionName);
@@ -100,47 +91,18 @@ public class LeadControllerManager {
   }
 
   /**
-   * Checks if the current controller host is Helix leader.
+   * Checks if the current controller host is Helix cluster leader.
    */
   private boolean isHelixLeader() {
     return _amIHelixLeader;
   }
 
   /**
-   * Response class to return lead controller id and partition index.
+   * Callback on changes in the controller. Should be registered to the controller callback. This callback is not needed when the resource is enabled.
+   * However, the resource can be disabled sometime while the cluster is in operation, so we keep it here. Plus, it does not add much overhead.
+   * At some point in future when we stop supporting the disabled resource, we will remove this line altogether and the logic that goes with it.
    */
-  public class LeadControllerResponse {
-    private int _partitionIndex;
-    private String _leadControllerId;
-
-    public LeadControllerResponse(int partitionIndex, String leadControllerId) {
-      _partitionIndex = partitionIndex;
-      _leadControllerId = leadControllerId;
-    }
-
-    public LeadControllerResponse() {
-      _partitionIndex = -1;
-      _leadControllerId = null;
-    }
-
-    public int getPartitionIndex() {
-      return _partitionIndex;
-    }
-
-    public String getLeadControllerId() {
-      return _leadControllerId;
-    }
-
-    @Override
-    public String toString() {
-      return "Partition index: " + getPartitionIndex() + ", lead controller id: " + getLeadControllerId();
-    }
-  }
-
-  /**
-   * Callback on changes in the controller. Should be registered to the controller callback.
-   */
-  public void onHelixControllerChange() {
+  void onHelixControllerChange() {
     if (_pinotHelixResourceManager.isHelixLeader()) {
       if (!_amIHelixLeader) {
         _amIHelixLeader = true;
@@ -156,8 +118,5 @@ public class LeadControllerManager {
         LOGGER.info("Already not Helix leader. Duplicate notification");
       }
     }
-  }
-
-  public void onLeadControllerChange() {
   }
 }
